Refactor: Streamline Favourite entity relationships in 'com.mpm.beforeandafter.user.model' package
Removed the favouriteUser field and its associated annotations as it was redundant.
Updated the @JoinColumn annotation on the user field to include nullable = false to enforce a non-null constraint at the database level.
This refactor simplifies the Favourite entity by eliminating unnecessary relationships and ensuring data integrity with a non-null constraint on the user relationship.
These changes help to maintain a cleaner and more understandable entity representation for managing user favourites, aligning with the application's data model requirements.