refactor: Improved User entity with new associations
This commit enhances the User entity by adding new associations to ContactDetails and Image entities, creating a more comprehensive representation of a user's data. Now, each user can be associated with their contact details and a list of images. This update provides better support for maintaining user-related information in the system.